New feature: passthrough of UI options and CSS to OpenPage() and visualize()
The LDMClass functions openPage() and AddElement() now allows a passthrough of two optional parameters: uioptions (object) and uicss (string)

When given (as named parameters) in the python script, they are passed through in the data object which is presented by the oobd.visualize() callback.

In the oobd.openpage() function instead, to be compatible with older implementations, the first parameter stays the page name as it was before, but the function now presents also a second parameter, which contains the whole data object which came from the backend and so this object then also contains the ui parameter, if given.